fix error when writing file with empty content
Previously, attempting to write a local file with content of "" would set the resource Id to "", since the Id is based on a SHA1 of the content. This caused the file to be deleted during the apply step.
The Id is now set to a constant "-" to make it clear that it is not significant.

The ForceNew behaviour for content has been replaced with an Update function for more appropriate semantics.